The actual diagram we use in class
Team S@S -

This is the actual diagram we use in class which differs from the one in the S@S guide in one important respect: it shows that PO's need to exist for Knowledge & Infrastructure virtual teams so that they have someone who can prioritize their speciality backlogs. Have found this a very valuable addition in the field.