Improved description of SoSM responsibility
"the distribution of backlog" is confusing because that is normally something that is done with the PO role. What is meant here is that backlog may be passed along due to an impediment like one team being overloaded. The description has been changed to reflect the purpose of enabling cross team support and that passing backlog is an example.