Update guide-us-english.tex
The part of the sentence that states “operating consistently with the Scrum Guide” seems superfluous to me, yet it makes the definition more difficult to understand. Clean clear definitions help with understanding.
I understand that many people do not implement Scrum consistently with the Scrum Guide, but that is what the Reference Model section is for. Let's try to make that section a bit more important instead of making definitions longer than they need to be