Update guide-us-english.tex
It is my strong opinion that a very big win for Scrum at Scale is its focus on removing impediments (therefore optimizing Process Efficiency). I would love to see addressed explicitly in the guide that the ability to do so is what makes this framework exceptionally good. (Better than others) 
Page 2 seems like the kind of page to do that. 

On top of that, the paragraph that refers to the separation of the what and the how at the end of page 2 should immediately follow the paragraph that introduced it.