Update guide-us-english.tex
I believe the current recommendation is to have the most important impediment at the very top of the backlog at the start of every sprint. Since Product Owners have a tendency to misunderstand the importance of removing impediments and solving technical debt, I feel that this should be worded a bit stronger than it is right now. 
On top of that, technical debt refers quite strongly to programming, I prefer removing it and focusing on the resolving of impediments