EAT works through organizational issues
In the sentence "This set of teams works through organizational issues that block agility and creates a Reference Model for Scrum that is known to work in the organization and can be used as a pattern for scaling Scrum across the organization." the usage of "this set of teams" may be a bit confusing. It may refer to (a) EAT (a set of individuals) or to (b) the set of teams that form the Reference Model (I guess they are rather focusing on development than on organizational issues). Therefore, I suggest using EAT.